Rewrite JSONP interface to use HTTP
Data upload and session configuration working, query/facet-selection not quite.

The changes are primarily to Unhosted mode, but we've turned it into a
form of hosting by requiring data upload, so this could do with a rename.
Hosted mode is still there - where all data ends up in a single Sesame
database - but the new Unhosted mode seems potentially more manageable.
Feedback welcome.